security start

+++ src/Controller/CatPremierController.php
@@ -9,40 +9,61 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+
 
 /**
  * @Route("/cat/premier")
  */
 class CatPremierController extends AbstractController
 {
+    
     /**
      * @Route("/", name="app_cat_premier_index", methods={"GET"})
      */
     public function index(CatPremierRepository $catPremierRepository): Response
     {
-        return $this->render('cat_premier/index.html.twig', [
-            'cat_premiers' => $catPremierRepository->findAll(),
-        ]);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+            return $this->render('cat_premier/index.html.twig', [
+                'cat_premiers' => $catPremierRepository->findAll(),
+            ]);
+        }
+
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+            return $this->redirectToRoute('home');
+        }
     }
+
 
     /**
      * @Route("/new", name="app_cat_premier_new", methods={"GET", "POST"})
      */
     public function new(Request $request, CatPremierRepository $catPremierRepository): Response
     {
-        $catPremier = new CatPremier();
-        $form = $this->createForm(CatPremierType::class, $catPremier);
-        $form->handleRequest($request);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $catPremierRepository->add($catPremier);
-            return $this->redirectToRoute('app_cat_premier_index', [], Response::HTTP_SEE_OTHER);
+            $catPremier = new CatPremier();
+            $form = $this->createForm(CatPremierType::class, $catPremier);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $catPremierRepository->add($catPremier);
+                return $this->redirectToRoute('app_cat_premier_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->renderForm('cat_premier/new.html.twig', [
+                'cat_premier' => $catPremier,
+                'form' => $form,
+            ]);
         }
-
-        return $this->renderForm('cat_premier/new.html.twig', [
-            'cat_premier' => $catPremier,
-            'form' => $form,
-        ]);
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -50,9 +71,18 @@ class CatPremierController extends AbstractController
      */
     public function show(CatPremier $catPremier): Response
     {
-        return $this->render('cat_premier/show.html.twig', [
-            'cat_premier' => $catPremier,
-        ]);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+            return $this->render('cat_premier/show.html.twig', [
+                'cat_premier' => $catPremier,
+            ]);
+        }
+
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -60,18 +90,27 @@ class CatPremierController extends AbstractController
      */
     public function edit(Request $request, CatPremier $catPremier, CatPremierRepository $catPremierRepository): Response
     {
-        $form = $this->createForm(CatPremierType::class, $catPremier);
-        $form->handleRequest($request);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $catPremierRepository->add($catPremier);
-            return $this->redirectToRoute('app_cat_premier_index', [], Response::HTTP_SEE_OTHER);
+            $form = $this->createForm(CatPremierType::class, $catPremier);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $catPremierRepository->add($catPremier);
+                return $this->redirectToRoute('app_cat_premier_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->renderForm('cat_premier/edit.html.twig', [
+                'cat_premier' => $catPremier,
+                'form' => $form,
+            ]);
         }
 
-        return $this->renderForm('cat_premier/edit.html.twig', [
-            'cat_premier' => $catPremier,
-            'form' => $form,
-        ]);
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -79,10 +118,21 @@ class CatPremierController extends AbstractController
      */
     public function delete(Request $request, CatPremier $catPremier, CatPremierRepository $catPremierRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$catPremier->getId(), $request->request->get('_token'))) {
-            $catPremierRepository->remove($catPremier);
-        }
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        return $this->redirectToRoute('app_cat_premier_index', [], Response::HTTP_SEE_OTHER);
+            if ($this->isCsrfTokenValid('delete'.$catPremier->getId(), $request->request->get('_token'))) {
+                $catPremierRepository->remove($catPremier);
+            }
+
+            return $this->redirectToRoute('app_cat_premier_index', [], Response::HTTP_SEE_OTHER);
+        }
+        
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+            return $this->redirectToRoute('home');
+        }
     }
 }
+
+
