+++ src/Controller/CategoryBlogController.php
@@ -5,10 +5,11 @@ namespace App\Controller;
 use App\Entity\CategoryBlog;
 use App\Form\CategoryBlogType;
 use App\Repository\CategoryBlogRepository;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
  * @Route("/category/blog")
@@ -20,39 +21,67 @@ class CategoryBlogController extends AbstractController
      */
     public function index(CategoryBlogRepository $categoryBlogRepository): Response
     {
-        return $this->render('category_blog/index.html.twig', [
-            'category_blogs' => $categoryBlogRepository->findAll(),
-        ]);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+            return $this->render('category_blog/index.html.twig', [
+                'category_blogs' => $categoryBlogRepository->findAll(),
+            ]);
+        }
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+		    return $this->redirectToRoute('home');
+        }
     }
+
 
     /**
      * @Route("/new", name="app_category_blog_new", methods={"GET", "POST"})
      */
     public function new(Request $request, CategoryBlogRepository $categoryBlogRepository): Response
     {
-        $categoryBlog = new CategoryBlog();
-        $form = $this->createForm(CategoryBlogType::class, $categoryBlog);
-        $form->handleRequest($request);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $categoryBlogRepository->add($categoryBlog);
-            return $this->redirectToRoute('app_category_blog_index', [], Response::HTTP_SEE_OTHER);
+            $categoryBlog = new CategoryBlog();
+            $form = $this->createForm(CategoryBlogType::class, $categoryBlog);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $categoryBlogRepository->add($categoryBlog);
+                return $this->redirectToRoute('app_category_blog_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->renderForm('category_blog/new.html.twig', [
+                'category_blog' => $categoryBlog,
+                'form' => $form,
+            ]);
         }
 
-        return $this->renderForm('category_blog/new.html.twig', [
-            'category_blog' => $categoryBlog,
-            'form' => $form,
-        ]);
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+		    return $this->redirectToRoute('home');
+        }
     }
+
 
     /**
      * @Route("/{id}", name="app_category_blog_show", methods={"GET"})
      */
     public function show(CategoryBlog $categoryBlog): Response
     {
-        return $this->render('category_blog/show.html.twig', [
-            'category_blog' => $categoryBlog,
-        ]);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+            return $this->render('category_blog/show.html.twig', [
+                'category_blog' => $categoryBlog,
+            ]);
+        }
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+		    return $this->redirectToRoute('home');
+        }
+
     }
 
     /**
@@ -60,29 +89,46 @@ class CategoryBlogController extends AbstractController
      */
     public function edit(Request $request, CategoryBlog $categoryBlog, CategoryBlogRepository $categoryBlogRepository): Response
     {
-        $form = $this->createForm(CategoryBlogType::class, $categoryBlog);
-        $form->handleRequest($request);
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $categoryBlogRepository->add($categoryBlog);
-            return $this->redirectToRoute('app_category_blog_index', [], Response::HTTP_SEE_OTHER);
+            $form = $this->createForm(CategoryBlogType::class, $categoryBlog);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $categoryBlogRepository->add($categoryBlog);
+                return $this->redirectToRoute('app_category_blog_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->renderForm('category_blog/edit.html.twig', [
+                'category_blog' => $categoryBlog,
+                'form' => $form,
+            ]);
         }
 
-        return $this->renderForm('category_blog/edit.html.twig', [
-            'category_blog' => $categoryBlog,
-            'form' => $form,
-        ]);
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+            return $this->redirectToRoute('home');
+        }
     }
+
 
     /**
      * @Route("/{id}", name="app_category_blog_delete", methods={"POST"})
      */
     public function delete(Request $request, CategoryBlog $categoryBlog, CategoryBlogRepository $categoryBlogRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$categoryBlog->getId(), $request->request->get('_token'))) {
-            $categoryBlogRepository->remove($categoryBlog);
-        }
+        try {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        return $this->redirectToRoute('app_category_blog_index', [], Response::HTTP_SEE_OTHER);
+            if ($this->isCsrfTokenValid('delete'.$categoryBlog->getId(), $request->request->get('_token'))) {
+                $categoryBlogRepository->remove($categoryBlog);
+            }  
+            return $this->redirectToRoute('app_category_blog_index', [], Response::HTTP_SEE_OTHER);
+        }
+        catch(AccessDeniedException $ex){
+            $this->addFlash('error', "Vous n'avez pas les droits necessaires pour accèder à cette fonction");
+            return $this->redirectToRoute('home');
+        }
     }
 }
